% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CV.Zresidual.coxph.frailty.R
\name{CV.Zresidual.coxph.frailty}
\alias{CV.Zresidual.coxph.frailty}
\title{CV Zresidual for shared fraitly model using coxph function}
\usage{
CV.Zresidual.coxph.frailty(fit.coxph, data, nfolds, foldlist, n.rep)
}
\description{
CV Zresidual for shared fraitly model using coxph function
}
