% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dhurdle.nb.R
\name{dhurdle.nb}
\alias{dhurdle.nb}
\title{A function to calculate pdf of hurdle negative binomial.}
\usage{
dhurdle.nb(y, mu, size, pi, log = FALSE)
}
\description{
A function to calculate pdf of hurdle negative binomial.
}
