% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dhurdle.poisson.R
\name{dhurdle.pois}
\alias{dhurdle.pois}
\title{A function to calculate pdf of hurdle poisson.}
\usage{
dhurdle.pois(y, lambda, pi, log = FALSE)
}
\description{
A function to calculate pdf of hurdle poisson.
}
