% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dhurdle.poisson.R
\name{dhurdle.poisson}
\alias{dhurdle.poisson}
\title{A function to calculate pdf of hurdle poisson.}
\usage{
dhurdle.poisson(y, mu, pi, log = FALSE)
}
\description{
A function to calculate pdf of hurdle poisson.
}
