% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior.pred.R
\docType{data}
\name{.LINK_FUNCTIONS}
\alias{.LINK_FUNCTIONS}
\title{A function to calculate posterior parameter estimates}
\format{
An object of class \code{list} of length 13.
}
\usage{
.LINK_FUNCTIONS
}
\arguments{
\item{fit}{A `brm` fit.}

\item{dpar}{parameter.}
}
\description{
A function to calculate posterior parameter estimates
}
\keyword{datasets}
