% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iscv_logrpp.R
\name{iscv_logrpp}
\alias{iscv_logrpp}
\title{A function to calculate important sampling cross validation log rpp}
\usage{
iscv_logrpp(log_cdf, log_pmf)
}
\arguments{
\item{log_cdf}{log cdf.}

\item{log_pmf}{log pmf.}
}
\description{
A function to calculate important sampling cross validation log rpp
}
