% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legend.box.R
\name{legend.box}
\alias{legend.box}
\title{A function to draw qq plot of a z-residual}
\usage{
legend.box(legend.args)
}
\arguments{
\item{legend.args}{args}
}
\description{
A function to draw qq plot of a z-residual
}
