% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.pred.dist.HNB.R
\name{log.pred.dist.HNB}
\alias{log.pred.dist.HNB}
\title{A function to calculate predictive log hurdle negative binomial pmf and cdf of a 'brm' fit}
\usage{
\method{log}{pred.dist.HNB}(fit)
}
\arguments{
\item{fit}{A `brm` fit.}
}
\description{
A function to calculate predictive log hurdle negative binomial pmf and cdf of a 'brm' fit
}
