% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.pred.dist.NB.R
\name{log.pred.dist.NB}
\alias{log.pred.dist.NB}
\title{A function to calculate predictive log negative binomial pmf and cdf of 'brm' fit.}
\usage{
\method{log}{pred.dist.NB}(fit)
}
\arguments{
\item{fit}{A `brm` fit.}
}
\description{
A function to calculate predictive log negative binomial pmf and cdf of 'brm' fit.
}
