% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.pred.dist.pois.R
\name{log.pred.dist.pois}
\alias{log.pred.dist.pois}
\title{A function to calculate predictive log poisson pmf and cdf of 'brm' fit.}
\usage{
\method{log}{pred.dist.pois}(fit)
}
\arguments{
\item{fit}{A `brm` fit.}
}
\description{
A function to calculate predictive log poisson pmf and cdf of 'brm' fit.
}
