% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.pred.dist.TNB.R
\name{log.pred.dist.TNB}
\alias{log.pred.dist.TNB}
\title{A function to calculate predictive log truncated negative binomial pmf and cdf of 'brm' fit.}
\usage{
\method{log}{pred.dist.TNB}(fit)
}
\arguments{
\item{fit}{A `brm` fit.}
}
\description{
A function to calculate predictive log truncated negative binomial pmf and cdf of 'brm' fit.
}
