% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.pred.dist.zero.R
\name{log.pred.dist.zero}
\alias{log.pred.dist.zero}
\title{A function to calculate log predictive distribution (pmf and cdf) of logistic component of a 'brm' fit}
\usage{
\method{log}{pred.dist.zero}(fit)
}
\arguments{
\item{fit}{A `brm` fit.}
}
\description{
A function to calculate log predictive distribution (pmf and cdf) of logistic component of a 'brm' fit
}
