% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_diff_exp.R
\name{log_diff_exp}
\alias{log_diff_exp}
\title{A function to compute the logarithm of the difference between the exponentials of two log values.}
\usage{
log_diff_exp(la, lb)
}
\arguments{
\item{la}{A log value.}

\item{lb}{A log value.}
}
\description{
A function to compute the logarithm of the difference between the exponentials of two log values.
}
