% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_pointpred_survival_coxph_frailty.R
\name{log_pointpred_survival_coxph.penal}
\alias{log_pointpred_survival_coxph.penal}
\title{Predictive quantities for survival::coxph with frailty}
\usage{
log_pointpred_survival_coxph.penal(fit, data, traindata, ...)
}
\arguments{
\item{fit}{A fitted coxph object, usually inheriting from coxph.penal.}

\item{data}{New data to evaluate.}

\item{traindata}{Training data used to fit the model. This is required for
reconstructing the baseline cumulative hazard.}

\item{...}{Additional arguments passed from Zresidual.}
}
\value{
A list containing log-survival probabilities, log-likelihood
contributions, discreteness indicators, and linear predictors.
}
\description{
Predictive quantities for survival::coxph with frailty
}
