% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_pred_dist_zero.R
\name{log_pred_dist_zero}
\alias{log_pred_dist_zero}
\title{A function to calculate log predictive distribution (pmf and cdf) of logistic component of a 'brm' fit}
\usage{
log_pred_dist_zero(fit)
}
\arguments{
\item{fit}{A `brm` fit.}
}
\description{
A function to calculate log predictive distribution (pmf and cdf) of logistic component of a 'brm' fit
}
