% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.matrix.coxph.R
\name{model.frame.coxph}
\alias{model.frame.coxph}
\title{An internal S3 method for model.matrix.}
\usage{
model.frame.coxph(formula, ...)
}
\description{
An internal S3 method for model.matrix.
}
\keyword{internal}
