% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.frame.survreg.R
\name{model.frame.survreg}
\alias{model.frame.survreg}
\title{An internal S3 method for model.matrix.}
\usage{
\method{model.frame}{survreg}(formula, ...)
}
\description{
An internal S3 method for model.matrix.
}
\keyword{internal}
