% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.matrix.coxph.R
\name{model.matrix.coxph}
\alias{model.matrix.coxph}
\title{An internal S3 method for model.matrix.}
\usage{
model.matrix.coxph(object, data = NULL, contrast.arg = object$contrasts, ...)
}
\description{
An internal S3 method for model.matrix.
}
\keyword{internal}
