% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.frame.survreg.R
\name{model.matrix.survreg}
\alias{model.matrix.survreg}
\title{An internal S3 method for model.matrix.}
\usage{
model.matrix.survreg(object, data, ...)
}
\description{
An internal S3 method for model.matrix.
}
\keyword{internal}
