% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phurdle.nb.R
\name{phurdle.nb}
\alias{phurdle.nb}
\title{A function to calculate predictive probability of hurdle negative binomial.}
\usage{
phurdle.nb(y, mu, size, pi, lower.tail = FALSE, log.p = FALSE)
}
\arguments{
\item{y}{y values.}

\item{mu}{mu parameter.}

\item{size}{parameter.}

\item{pi}{hu parameter.}
}
\description{
A function to calculate predictive probability of hurdle negative binomial.
}
