% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_logmpp.R
\name{post_logmpp}
\alias{post_logmpp}
\title{Compute Posterior Log Middle-value Predictive p-values (Log-MPP)}
\usage{
post_logmpp(log_cdf, log_pmf)
}
\arguments{
\item{log_cdf}{Numeric matrix of log cumulative probabilities for each observation
(posterior samples × observations).}

\item{log_pmf}{Numeric matrix of log probability mass function values for each observation
(posterior samples × observations).}
}
\value{
A numeric vector of log posterior mid-value predictive probabilities for each observation.
}
\description{
Calculates the posterior log middle-value (0.5) predictive p-values (log-MPP) for a set
of observations given the log-PMF and log-CDF matrices obtained from posterior
predictive samples.
}
\details{
The function computes a stabilized version of the posterior predictive p-values:
\enumerate{
\item It sums the log-CDF and log-PMF across posterior samples using log-sum-exp for numerical stability.
\item Probabilities exactly equal to 0 or 1 are replaced with small bounds (\eqn{10^{-5}} and \eqn{9 \cdot 10^{-5}}) to avoid numerical issues in further computations.
}
}
\examples{
NULL

}
