% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior.pred.R
\name{posterior.pred}
\alias{posterior.pred}
\title{A function to calculate posterior parameter estimates}
\usage{
posterior.pred(fit, dpar, count.only = TRUE)
}
\arguments{
\item{fit}{A `brm` fit.}

\item{dpar}{parameter.}
}
\description{
A function to calculate posterior parameter estimates
}
