% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predcheck_top.R
\name{print.moment_predcheck}
\alias{print.moment_predcheck}
\title{Print a moment predictive-check summary}
\usage{
\method{print}{moment_predcheck}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"moment_predcheck"}.}

\item{...}{Reserved for future use.}
}
\value{
The input object, invisibly.
}
\description{
Print a moment predictive-check summary
}
