% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predcheck_top.R
\name{print.predcheck}
\alias{print.predcheck}
\title{Print a predictive-check summary}
\usage{
\method{print}{predcheck}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"predcheck"}.}

\item{...}{Reserved for future use.}
}
\value{
The input object, invisibly.
}
\description{
Print a predictive-check summary
}
