% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.residuals.R
\name{print.resid}
\alias{print.resid}
\title{Print a residual vector}
\usage{
\method{print}{resid}(x, ...)
}
\arguments{
\item{x}{A residual object or numeric vector.}

\item{...}{Further arguments passed to \code{print()}.}
}
\value{
The input object, invisibly.
}
\description{
Simple print method for residual-like vectors used in this package.
}
\examples{
x <- structure(rnorm(5), class = "resid")
print(x)

}
