% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predcheck_top.R
\name{print.z_predcheck}
\alias{print.z_predcheck}
\title{Print a Z-residual predictive-check summary}
\usage{
\method{print}{z_predcheck}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"z_predcheck"}.}

\item{...}{Reserved for future use.}
}
\value{
The input object, invisibly.
}
\description{
Print a Z-residual predictive-check summary
}
