% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmin.R
\name{pvalue.min}
\alias{pvalue.min}
\title{A function to calculate the min p-value}
\usage{
pvalue.min(pv)
}
\arguments{
\item{pv}{Numeric vector of p-values.}
}
\value{
A single numeric value: the minimum adjusted p-value.
}
\description{
A function to calculate the min p-value
}
