% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuals.R
\name{residuals}
\alias{residuals}
\title{A function to calculate other type of residual}
\usage{
residuals(
  fit.object,
  data,
  residual.type = c("censored Z-residual", "Cox-Snell", "martingale", "deviance")
)
}
\value{
\itemize{
 \item{censored.Zresid}{censored Z-residual}
 \item{ucs}{unmodified CS residuals}
 \item{martg}{Martingale residuals}
 \item{dev}{Deviance residuals}
}
}
\description{
A function to calculate other type of residual
}
