% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanitize_names.r
\name{sanitize_names}
\alias{sanitize_names}
\title{Sanitize column names (internal)}
\usage{
sanitize_names(x)
}
\arguments{
\item{x}{Character vector of names.}
}
\value{
Character vector with sanitized names.
}
\description{
Replace problematic characters in column names
to get something closer to brms' naming scheme.
}
\keyword{internal}
