% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf.test.zresid.R
\name{sf.test.zresid}
\alias{sf.test.zresid}
\title{A function to calculate Shapiro-Francia test of Zresidual}
\usage{
sf.test.zresid(Zresidual)
}
\arguments{
\item{Zresidual}{A Z-residual.}
}
\description{
A function to calculate Shapiro-Francia test of Zresidual
}
