% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bartlett.test.zresid.R
\name{test.var.bartl}
\alias{test.var.bartl}
\title{Bartlett test core (internal): variance homogeneity across bins/levels}
\usage{
test.var.bartl(Zresidual, fitted.value, k.bl = 10)
}
\arguments{
\item{Zresidual}{numeric vector}

\item{fitted.value}{numeric/factor vector (same length as Zresidual)}

\item{k.bl}{number of bins for numeric covariates}
}
\value{
p-value (numeric) or NA_real_
}
\description{
Bartlett test core (internal): variance homogeneity across bins/levels
}
\keyword{internal}
