% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.Zresidual.R
\docType{package}
\name{Zresidual-package}
\alias{Zresidual-package}
\title{Zresidual: Diagnostics with Z-residuals}
\description{
This package provides functions for computing and analyzing Z-residuals and cross-validatory Z-residuals from fitting outputs of several R packages, including coxph, survreg, glmmTMB, brms, ...
}
\seealso{
Useful links:
\itemize{
  \item \url{https://tiw150.github.io/Zresidual}
  \item Report bugs at \url{https://github.com/tiw150/Zresidual/issues}
}

}
\author{
\strong{Maintainer}: Tingxuan Wu \email{tiw150@mail.usask.ca}

Authors:
\itemize{
  \item Longhai Li \email{longhai@math.usask.ca}
}

Other contributors:
\itemize{
  \item Hettiarachchige Alias Egodage Dananji \email{ocf542@mail.usask.ca} [contributor]
  \item Cindy Feng \email{longhai@math.usask.ca} [contributor]
}

}
\keyword{internal}
