% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Zresidual.coxph.frailty.R
\name{Zresidual.coxph.frailty}
\alias{Zresidual.coxph.frailty}
\title{A title}
\usage{
Zresidual.coxph.frailty(fit_coxph, traindata, newdata, n.rep = 1)
}
\arguments{
\item{fit_coxph}{is the fit object are from the coxph function with frailty item in the survival package.}
}
\value{
\itemize{
 \item{Zresid}{Z-residual}
}
}
\description{
Some descriptions
}
