% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Zresidual.coxph.R
\name{Zresidual.coxph}
\alias{Zresidual.coxph}
\title{A title}
\usage{
Zresidual.coxph(fit_coxph, newdata, n.rep = 1)
}
\arguments{
\item{fit_coxph}{is the fit object are from the coxph function in the survival package.}
}
\value{
\itemize{
 \item{Zresid}{Z-residual}
}
}
\description{
Some descriptions
}
