% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Zresidual.hurdle.poisson.R
\name{Zresidual.hurdle.poisson}
\alias{Zresidual.hurdle.poisson}
\title{A function to calculate z-residuals of a 'brm' fit.
This function is to be used when the user needs to calculate the Z-residuals of TP/HP}
\usage{
Zresidual.hurdle.poisson(fit, type, method = "iscv", n.rep = 1)
}
\arguments{
\item{fit}{is the fit object are from 'brms' package.}
}
\value{
\itemize{
 \item{Zresid} {Z-residual}
 }
}
\description{
A function to calculate z-residuals of a 'brm' fit.
This function is to be used when the user needs to calculate the Z-residuals of TP/HP
}
