% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Zresidual.survreg.R
\name{Zresidual.survreg}
\alias{Zresidual.survreg}
\title{Zresidual for survival}
\usage{
Zresidual.survreg(survreg_fit, newdata)
}
\arguments{
\item{survreg_fit}{a survreg object}
}
\description{
Zresidual for survival
}
