% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Zresidual_bernoulli.R
\name{zresidual_bernoulli}
\alias{zresidual_bernoulli}
\title{A function to calculate z-residuals of a 'brm' fit.
This function is to be used when the user needs to calculate the z-residuals of bernoulli}
\usage{
zresidual_bernoulli(fit, method = "iscv", nrep = 1)
}
\description{
A function to calculate z-residuals of a 'brm' fit.
This function is to be used when the user needs to calculate the z-residuals of bernoulli
}
