% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Zresidual.coxph.survival.R
\name{Zresidual_coxph_frailty_survival}
\alias{Zresidual_coxph_frailty_survival}
\title{Z-residuals for Standard Cox Models with Frailty terms (Internal Worker)}
\usage{
Zresidual_coxph_frailty_survival(fit_coxph, traindata, newdata, n.rep = 1, ...)
}
\description{
Z-residuals for Standard Cox Models with Frailty terms (Internal Worker)
}
\keyword{internal}
