% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Zresidual_hurdle_negbinomial.R
\name{zresidual_hurdle_negbinomial}
\alias{zresidual_hurdle_negbinomial}
\title{A function to calculate z-residuals of a 'brm' fit.
This function is to be used when the user needs to calculate the z-residuals of TNB/HNB}
\usage{
zresidual_hurdle_negbinomial(fit, type, method = "iscv", nrep = 1)
}
\description{
A function to calculate z-residuals of a 'brm' fit.
This function is to be used when the user needs to calculate the z-residuals of TNB/HNB
}
