% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Zresidual_hurdle_poisson.R
\name{zresidual_hurdle_poisson}
\alias{zresidual_hurdle_poisson}
\title{A function to calculate z-residuals of a 'brm' fit.
This function is to be used when the user needs to calculate the z-residuals of TP/HP}
\usage{
zresidual_hurdle_poisson(fit, type, method = "iscv", nrep = 1)
}
\description{
A function to calculate z-residuals of a 'brm' fit.
This function is to be used when the user needs to calculate the z-residuals of TP/HP
}
