% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Zresidual_negbinomial.R
\name{zresidual_negbinomial}
\alias{zresidual_negbinomial}
\title{A function to calculate z-residuals of a 'brm' fit.
This function is to be used when the user needs to calculate the z-residuals of TNB/HNB}
\usage{
zresidual_negbinomial(fit, type = "NB", method = "iscv", nrep = 1)
}
\description{
A function to calculate z-residuals of a 'brm' fit.
This function is to be used when the user needs to calculate the z-residuals of TNB/HNB
}
