% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aov.test.zresid.R
\name{aov.test.zresid}
\alias{aov.test.zresid}
\title{A function to calculate ANOVA of Zresidual}
\usage{
aov.test.zresid(Zresidual, X = c("lp", "covariate"), k.anova = 10)
}
\arguments{
\item{Zresidual}{A Z-residual.}

\item{X}{Linear predictor or covariate}

\item{k.anova}{Number of bins if applicable}
}
\description{
A function to calculate ANOVA of Zresidual
}
