% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bartlett.test.zresid.R
\name{bartlett.test.zresid}
\alias{bartlett.test.zresid}
\title{#' A function to calculate Bartlett of Zresidual}
\usage{
\method{bartlett.test}{zresid}(Zresidual, X = c("lp", "covariate"), k.bl = 10)
}
\arguments{
\item{Zresidual}{A z-residual.}

\item{X}{Linear predictor or covariate}

\item{k.bl}{Number of bins if applicable}
}
\description{
#' A function to calculate Bartlett of Zresidual
}
