% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplot.zresid.R
\name{boxplot.zresid}
\alias{boxplot.zresid}
\title{A function to draw box plot of a z-residual}
\usage{
\method{boxplot}{zresid}(
  Zresidual,
  irep = 1,
  X = c("lp", "covariate"),
  num.bin = 10,
  main.title = paste("Z-residual Boxplot -", attr(Zresidual, "type")),
  outlier.return = FALSE,
  outlier.value = 3.5,
  ...
)
}
\arguments{
\item{Zresidual}{A Z-residual.}
}
\description{
A function to draw box plot of a z-residual
}
