% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CV.Zresidual.R
\name{CV.Zresidual}
\alias{CV.Zresidual}
\title{A title}
\usage{
CV.Zresidual(fit.object, nfolds, foldlist = NULL, data = NULL, nrep = 1)
}
\arguments{
\item{fit.object}{The fit object are one of 'coxph' and 'survreg'.}

\item{data}{Data that used for fitting the survival model.}
}
\value{
\itemize{
 \item{Zresid}{CV Z-residual}
 \item{SP}{Survival Probabilities}
}
}
\description{
Some descriptions
}
