% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CV.Zresidual.survreg.R
\name{CV.Zresidual.survreg}
\alias{CV.Zresidual.survreg}
\title{CV Zresidual for shared fraitly model using coxph function}
\usage{
CV.Zresidual.survreg(fit.survreg, data, nfolds, foldlist, n.rep)
}
\description{
CV Zresidual for shared fraitly model using coxph function
}
