% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gof.censored.zresidual.R
\name{gof.censored.zresidual}
\alias{gof.censored.zresidual}
\title{A function to perform a goodness-of-fit test on censored Z-residuals.}
\usage{
gof.censored.zresidual(censored.Zresidual)
}
\arguments{
\item{censored.Zresidual}{The censored Z-residuals to be tested.}
}
\description{
A function to perform a goodness-of-fit test on censored Z-residuals.
}
