% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.pred.dist.TP.R
\name{log.pred.dist.TP}
\alias{log.pred.dist.TP}
\title{A function to calculate predictive log truncated poisson pmf and cdf of 'brm' fit.}
\usage{
\method{log}{pred.dist.TP}(fit)
}
\arguments{
\item{fit}{A `brm` fit.}
}
\description{
A function to calculate predictive log truncated poisson pmf and cdf of 'brm' fit.
}
