% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_pointpred_survival_coxph_frailty.R
\name{log_pointpred_survival_coxph.penal}
\alias{log_pointpred_survival_coxph.penal}
\title{Predictive quantities for survival::coxph with frailty}
\usage{
log_pointpred_survival_coxph.penal(fit, data, traindata, ...)
}
\arguments{
\item{fit}{A fitted coxph object (usually class coxph.penal).}

\item{data}{The new data to evaluate.}

\item{traindata}{The original data used to fit the model (required for baseline hazard).}

\item{...}{Additional arguments passed from Zresidual.}
}
\description{
Predictive quantities for survival::coxph with frailty
}
