% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf.tp.R
\name{pdf.tp}
\alias{pdf.tp}
\title{A function to calculate pmf of Truncated Poisson}
\usage{
pdf.tp(y, lambda, log.p = FALSE)
}
\arguments{
\item{y}{y values.}

\item{lambda}{lambda parameter of TP distribution.}
}
\description{
A function to calculate pmf of Truncated Poisson
}
