% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.cs.residual.R
\name{plot.cs.residual}
\alias{plot.cs.residual}
\title{A function to draw scatter plot of a Cox-snell Residual}
\usage{
\method{plot}{cs.residual}(
  cs.residual,
  ylab = "Cumulative Hazard Function",
  main.title = "Cox-Snell Residuals Scatterplot",
  outlier.return = FALSE,
  ...
)
}
\arguments{
\item{cs.residual}{A numeric vector of Cox-Snell residuals.}
}
\description{
A function to draw scatter plot of a Cox-snell Residual
}
