% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.dev.resid.R
\name{plot.dev.resid}
\alias{plot.dev.resid}
\title{A function to draw scatter plot of a Deviance Residual}
\usage{
\method{plot}{dev.resid}(
  Deviance.residual,
  ylab = "Deviance Residual",
  X = c("index", "lp", "covariate"),
  main.title = "Deviance Residual Plot",
  outlier.return = FALSE,
  ...
)
}
\arguments{
\item{Deviance.residual}{}
}
\description{
A function to draw scatter plot of a Deviance Residual
}
