% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.martg.resid.R
\name{plot.martg.resid}
\alias{plot.martg.resid}
\title{A function to draw scatter plot of a Martingale Residual}
\usage{
\method{plot}{martg.resid}(
  Martingale.residual,
  ylab = "Martingale Residual",
  X = c("index", "lp", "covariate"),
  main.title = "Martingale Residual Plot",
  outlier.return = FALSE,
  ...
)
}
\arguments{
\item{Martingale.residual}{}
}
\description{
A function to draw scatter plot of a Martingale Residual
}
