% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_logrpp.R
\name{post_logrpp}
\alias{post_logrpp}
\title{A function to calculate posterior log rpp}
\usage{
post_logrpp(log_cdf, log_pmf)
}
\arguments{
\item{log_cdf}{log cdf.}

\item{log_pmf}{log pmf.}
}
\description{
A function to calculate posterior log rpp
}
