% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qqnorm.zresid.R
\name{qqnorm.zresid}
\alias{qqnorm.zresid}
\title{A function to draw qq plot of a z-residual}
\usage{
\method{qqnorm}{zresid}(
  Zresidual,
  irep = 1,
  diagnosis.test = c("SW", "ANOVA"),
  X.anova = c("lp", "covariate"),
  k.anova = 10,
  main.title = ifelse(is.null(attr(Zresidual, "type")), "Normal Q-Q Plot",
    paste("Normal Q-Q Plot -", attr(Zresidual, "type"))),
  xlab = "Theoretical Quantiles",
  ylab = "Sample Quantiles",
  outlier.return = TRUE,
  outlier.value = 3.5,
  outlier.set = list(),
  legend.settings = list(),
  ...
)
}
\arguments{
\item{Zresidual}{A z-residual.}

\item{diagnosis.test}{diagnosis test}

\item{X.anova}{X variable to run if ANOVA is selected in diagnosis.test. Possible c("lp", "covariate")}

\item{k.anova}{Number of bins if X.anova is 'covariates'.}

\item{legend.settings}{A list of parameters available in legend().}

\item{point.clr}{A vector of colors of points.}

\item{point.type}{A vector of type of points.}

\item{outlier.settings}{A list of parameters available in symbols() and text().}
}
\description{
A function to draw qq plot of a z-residual
}
