% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sw.test.zresid.R
\name{sw.test.zresid}
\alias{sw.test.zresid}
\title{A function to calculate Shapiro Wilk test of Zresidual}
\usage{
sw.test.zresid(Zresidual, ...)
}
\arguments{
\item{Zresidual}{A Z-residual.}
}
\description{
A function to calculate Shapiro Wilk test of Zresidual
}
