% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.nl.aov.R
\name{test.nl.aov}
\alias{test.nl.aov}
\title{A function to calculate ANOVA}
\usage{
test.nl.aov(Zresidual, fitted.value, k.anova = 10)
}
\arguments{
\item{Zresidual}{A Z-residual.}

\item{fitted.value}{Fitted values}

\item{k.anova}{Number of bins if applicable}
}
\description{
A function to calculate ANOVA
}
