% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.var.bartl.R
\name{test.var.bartl}
\alias{test.var.bartl}
\title{A function to calculate Bartlett test of Zresidual}
\usage{
test.var.bartl(Zresidual, fitted.value, k.bl = 10)
}
\arguments{
\item{Zresidual}{A Z-residual.}

\item{fitted.value}{Fitted values}

\item{k.bl}{Number of bins if applicable}
}
\description{
A function to calculate Bartlett test of Zresidual
}
