% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Zcov.r
\name{Zcov}
\alias{Zcov}
\title{Zcov: extract covariates / lp / response / y_type for diagnostics}
\usage{
Zcov(
  fit,
  data = NULL,
  mode = c("auto", "hurdle_zero"),
  detail = c("basic", "full"),
  ...
)
}
\arguments{
\item{fit}{model object (brmsfit, coxph, survreg, ...)}

\item{data}{optional newdata}

\item{mode}{optional; default "auto"}

\item{detail}{optional; "basic" or "full"}
}
\description{
Zcov: extract covariates / lp / response / y_type for diagnostics
}
