% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.Zresidual.R
\docType{package}
\name{Zresidual-package}
\alias{Zresidual-package}
\title{Zresidual: Diagnostics with Z-residuals}
\description{
Provides a unified framework for computing and analyzing Z-residuals across a diverse range of statistical models. The package facilitates model diagnostics by transforming outputs from popular modeling frameworks—including 'coxph', 'survreg', 'glmmTMB', and 'brms'—into standardized Z-residuals. These tools enable researchers to assess model fit, identify outliers, and detect misspecification. In addition to standard residuals, the package implements cross-validatory variants to support robust predictive assessments and model validation workflows.URL: https://tiw150.github.io/Zresidual
}
\seealso{
Useful links:
\itemize{
  \item Report bugs at \url{https://github.com/tiw150/Zresidual/issues}
}

}
\author{
\strong{Maintainer}: Tingxuan Wu \email{tiw150@mail.usask.ca}

Authors:
\itemize{
  \item Longhai Li \email{longhai@math.usask.ca}
}

Other contributors:
\itemize{
  \item Hettiarachchige Alias Egodage Dananji \email{ocf542@mail.usask.ca} [contributor]
  \item Cindy Feng \email{longhai@math.usask.ca} [contributor]
}

}
\keyword{internal}
