% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Zresidual.bernoulli.R
\name{Zresidual.bernoulli}
\alias{Zresidual.bernoulli}
\title{A function to calculate z-residuals of a 'brm' fit.
This function is to be used when the user needs to calculate the z-residuals of bernoulli}
\usage{
Zresidual.bernoulli(fit, method = "iscv", nrep = 1)
}
\arguments{
\item{fit}{is the fit object are from 'brms' package.}
}
\value{

}
\description{
A function to calculate z-residuals of a 'brm' fit.
This function is to be used when the user needs to calculate the z-residuals of bernoulli
}
