% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Zresidual.poisson.R
\name{Zresidual.poisson}
\alias{Zresidual.poisson}
\title{A function to calculate z-residuals of a 'brm' fit.
This function is to be used when the user needs to calculate the Z-residuals of TNB/HNB}
\usage{
Zresidual.poisson(fit, type = "pois", method = "iscv", n.rep = 1)
}
\arguments{
\item{fit}{is the fit object are from 'brms' package.}
}
\value{
\itemize{
 \item{Zresid}{Z-residual}
 }
}
\description{
A function to calculate z-residuals of a 'brm' fit.
This function is to be used when the user needs to calculate the Z-residuals of TNB/HNB
}
