% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Zresidual.R
\name{Zresidual}
\alias{Zresidual}
\title{A title}
\usage{
Zresidual(fit.object, data, fit.object2 = NULL)
}
\arguments{
\item{fit.object}{The fit object are one of 'coxph', 'survreg' and 'glmmTMB'.}

\item{data}{Data that used for fitting the survival model.}
}
\value{
\itemize{
 \item{Zresid}{Z-residual}
 \item{censored.Zresid}{censored Z-residual}
 \item{SP}{Survival Probabilities}
 \item{ucs}{unmodified CS residuals}
 \item{mcs}{modified CS residuals}
 \item{martg}{Martingale residuals}
 \item{dev}{Deviance residuals}
}
}
\description{
Some descriptions
}
