% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Zresidual.survreg.R
\name{Zresidual.survreg}
\alias{Zresidual.survreg}
\title{Z-residuals for Accelerated failure time model models}
\usage{
Zresidual.survreg(fit_survreg, newdata, n.rep = 1)
}
\arguments{
\item{fit_survreg}{A fitted object from the `survreg` function in the `survival` package.}

\item{newdata}{Optional \code{data.frame} containing the variables used in \code{fit_coxph$formula}. If \code{NULL} (the default), residuals are computed on the original data used to fit the model. If supplied, \code{newdata} must contain the survival response and all covariates appearing in the original model formula.}

\item{n.rep}{An integer specifying the number of random draws to use for calculating the Z-residuals for censored observations. Defaults to `nrep` (which should be defined in your environment, a common choice is 100).}
}
\value{
A numeric matrix of dimension \eqn{n \times} \code{n.rep},  where \eqn{n} is the number of observations in the (new) model frame.   Each column corresponds to one set of Z-residuals. The returned matrix has the following attributes attached:
  \itemize{
    \item \code{Survival.Prob}: vector of survival probabilities \eqn{S_i(t_i)}.
    \item \code{linear.pred}: vector of linear predictors \eqn{\eta_i}.
    \item \code{covariates}: data frame of covariates (model frame without the response).
    \item \code{censored.status}: event indicator (1 = event,0 = censored).
    \item \code{object.model.frame}: the \code{model.frame} used to compute the residuals.
    \item \code{type}: character string \code{"survival"}.
  }
}
\description{
This function calculates Z-residuals based on a fitted `survreg` object from the `survival` package.
}
\examples{

library(survival)

# Fit a Weibull survival regression model
fit_weibull <- survreg(Surv(time, status) ~ age, data = lung, dist = "weibull")

# Calculate Z-residuals for the fitted model
z_residuals <- Zresidual.survreg(fit_weibull)
head(z_residuals)

# Calculate Z-residuals for new data
new_data <- data.frame(x = c(1, 2, 0.5))
Z_residuals_new <- Zresidual.survreg(fit_weibull, newdata = new_data)
head(z_residuals_new)


}
